extract utils to create service and change config style
