+++ ci/scripts/e2e-test-serial.sh
@@ -133,8 +133,8 @@ echo "--- e2e, $mode, Apache Superset"
 risedev slt -p 4566 -d dev './e2e_test/superset/*.slt' --junit "batch-${profile}"
 
 echo "--- e2e, $mode, embedding"
-cargo build --manifest-path e2e_test/vector_search/mocked_openai_embedding_service/Cargo.toml --locked --bin mocked_openai_embedding_service
-cargo run --manifest-path e2e_test/vector_search/mocked_openai_embedding_service/Cargo.toml --locked --bin mocked_openai_embedding_service &
+cargo build -p openai_embedding_service --bin mocked_service
+cargo run -p openai_embedding_service --bin mocked_service &
 MOCKED_EMBEDDING_SERVICE_PID=$!
 # wait for embedding service up
 sleep 3
