+++ src/expr/impl/Cargo.toml
@@ -18,7 +18,7 @@ udf = ["arrow-udf-runtime", "tonic", "ginepro", "zstd"]
 aho-corasick = "1"
 anyhow = "1"
 arrow-udf-runtime = { workspace = true, optional = true }
-async-openai = "0.29.0"
+async-openai = { workspace = true }
 async-trait = "0.1"
 auto_enums = { workspace = true }
 bytes = "1"
